Add custom icons building script (#166)
Adds a custom script based on svgstore and svgo that builds
icons in the same way it's done now.

The script supports configuration files like .designrc,
design.config.js and so on with help of `cosmiconfig`.

In the configuration file is possible to specify list
of images that should be built into the sprite and/or copied
to the project.

Adds icon mapping for the current icons `@mdi/svg` icons
in the default library's configuration file at the project
root (config.js). A few icons are different. CORE Logo is
removed from the sprite because it's not an icon.

Adds script dependencies:
- svgo (moved from devDependencies)
- svgstore

Removes current commands (build:svg:*) and svgstore-cli.

Adds "bin" section to the package.json and a single
bin command 'build-icons'. The command can be called from
the dependent projects.