Add smart Link component
Technically, <a> should have omitted some semantic attributes if there
is no href passed. The smart Link component solves this problem in
a the generic way. However, it pays a level more of the composition
for the this what's quite questionable solution when put
as default implementation.

Adds exporting both a smart BaseLink and StyledLink from the Link file.