Set icon size explicitly in the HTML
Prevents all the icons take all available space if the CSS was not
loaded accidentally, e.g. when the server fell down serving only HTML.

Could have unpredicted impact since requires both parameters:
`width` and `height` being overridden from CSS to change the size of
the icon actually.

Since the `Logo` is dependent on the `Icon`, it is updated too.